requirements.txt
fastapi>=0.119.1
Pillow>=12.0.0
torch>=2.7.0+cu126
torchvision>=0.22.0+cu126
